pyproject.toml
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.4,<4"]

[project]
authors = [{name = "Joel Payne", email = "lilspazjoekp@gmail.com"}]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Utilities"
]
dynamic = ["version", "description"]
name = "automodder"
requires-python = "~=3.11"
dependencies = [
  "praw",
  "click==8.*",
  "colorlog",
  "python-dateutil",
  "pydantic==2.*",
  "gspread",
  "gspread-formatting"
]

[project.optional-dependencies]
dev = [
  "pre-commit",
  "ruff ==0.1.*"
]

[project.scripts]
automodder = "automodder.cli:main"

[tool.black]
line-length = 88

[tool.isort]
profile = 'black'

[tool.ruff]
target-version = "py311"
include = [
  "automodder/**/*.py"
]
ignore = [
  "A002", # shadowing built-in
  "ANN101", # missing type annotation for self in method
  "ANN102", # missing type annotation for cls in classmethod
  "ANN202", # missing return type for private method
  "ANN401", # typing.Any usage
  "D203", # 1 blank line required before class docstring
  "D213", # Multi-line docstring summary should start at the second line
  "E501", # line-length
  "PLR0913", # too many arguments
  "PLR2004", # Magic value used in comparison,
  "S101" # use of assert
]
select = [
  "A", # flake8-builtins
  "ANN", # flake8-annotations
  "ARG", # flake8-unused-arguments
  "B", # flake8-bugbear
  "BLE", # flake8-blind-except
  "C4", # flake8-comprehensions
  "D", # pydocstyle
  "DTZ", # flake8-datetimez
  "E", # pycodestyle errors
  "EM", # flake8-errmsg
  "ERA", # eradicate
  "EXE", # flake8-executable
  "F", # pyflakes
  "FA", # flake8-future-annotations
  "FIX", # flake8-fix me
  "FLY", # flynt
  "G", # flake8-logging-format
  "I", # isort
  "INP", # flake8-no-pep420
  "ISC", # flake8-implicit-str-concat
  "N", # pep8-naming
  "PIE", # flake8-pie
  "PGH", # pygrep-hooks
  "PL", # Pylint
  "PT", # flake8-pytest-style
  "PTH", # flake8-use-pathlib
  "PYI", # flake8-pyi
  "Q", # flake8-quotes
  "RET", # flake8-return
  "RSE", # flake8-raise
  "S", # bandit
  "SIM", # flake8-simplify
  "T10", # flake8-debugger
  "T20", # flake8-print
  "TCH", # flake8-type-checking
  "TD", # flake8-todos
  "W", # pycodestyle warnings
  "UP" # pyupgrade
]
ignore-init-module-imports = true

[tool.ruff.flake8-annotations]
allow-star-arg-any = true
mypy-init-return = true
suppress-dummy-args = true
suppress-none-returning = true

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"automodder/logger.py" = ["G004"]
"automodder/managers/*.py" = ["PLE1205", "G004"]
